fully implented calculations for gear collector vision *UNTESTED*
Implements experimental calculations for comparing all visible targets and checking to see if any two are within the seperation and yDelta threshold to be a valid target for gear vision. After a valid set of targets are found, the x and y centroid of the two targets combined is found and a 3D vector is calculated and sent to the robot from that. This vector should refer exactly to where the peg is (in theory).

Things to do for this:
image_processor.cpp
* Measure target width and height threshold values
* Adjust fullness value

VisionTrackerGLSurfaceView.java
* Measure threshold values for seperation distance of the two targets
* Measure yDelta threshold values for the two targets