cleaned up code comments
Clarified some comments and set yDeltaMin value. yDeltaMin value shouldn't be necessary because the min for the delta in the y axis between the two centroids should logically always be 0. However, I'm leaving this variable in just in case we have some odd vision targets in future games that may require us to check for a minimum yDelta other than 0.