Gear vision almost working
After a LOT of trial and error (as well as research) I believe I know what needs to be done. We need to run this new version of the app on a phone while observing the output. The app will print the size of every target found, regardless of if it is accepted. So what we need to do is point the camera ONLY at one of the gear vision targets and observe what the printed poly.size is. Now go in to image_processor.cpp and modify the poly.size() == 4 statement to whatever the reported poly size is. After this more calibration may be neccassary for things like fullness and size. Additionally, the slope thresholds can be modified to require a stricter rectangle.